+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 ### Added
 
 - Added ASAN memory sanitization to QuickPool
+- Added File Memory Allocator
 
 ### Changed